+++ package.json
@@ -22,10 +22,10 @@
     "test:watch": "jest --watch",
     "test:cov": "jest --coverage",
     "test:debug": "node --inspect-brk -r tsconfig-paths/register -r ts-node/register node_modules/.bin/jest --runInBand",
-    "test:e2e": "jest --detectOpenHandles --maxWorkers=2  --config ./test/config/jest.config.e2e.js",
+    "test:e2e": "jest --forceExit --maxWorkers=4  --config ./test/config/jest.config.e2e.js",
     "test:it": "jest --forceExit --detectOpenHandles --isolatedModules=true --runInBand   --config ./test/config/jest.config.it.js",
     "test:ut": "jest --detectOpenHandles --runInBand --config ./test/config/jest.config.js",
-    "test:nightly": "jest --forceExit --maxWorkers=3 --config ./test/config/jest.config.nightly.js",
+    "test:nightly": "jest  --maxWorkers=4 --config ./test/config/jest.config.nightly.js",
     "test:ci": "jest --forceExit --runInBand --config ./test/config/jest.config.ci.js"
   },
   "dependencies": {
