+++ test/functional-e2e/common/messages-list.ts
@@ -1,0 +1,18 @@
+// Heading messages
+export const successMessageSignUp = 'Thank you for signing up!';
+export const successMessageVerifyEmail =
+  'Thank you for verifying your email address!';
+export const successAlert =
+  'An email containing a verification link has been sent to the email address you provided.';
+export const successMessageSentRecoveryMail =
+  'An email containing a recovery link has been sent to the email address you provided.';
+
+
+  // Alert messages
+  export const successMessageRecoveredAccount =
+  'You successfully recovered your account. Please change your password or set up an alternative login method (e.g. social sign in) within the next 15.00 minutes.';
+export const errorMessageEmptyField = 'Please fill required fields!';
+export const errorMessageSubmitRecoveryLinkTwice =
+  'Could not find a strategy to recover your account with. Did you fill out the form correctly?';
+export const errorMessageInvalidCredentials =
+  'The provided credentials are invalid, check for spelling mistakes in your password or username, email address, or phone number.';
