+++ test/functional-e2e/common/url-list.ts
@@ -1,0 +1,6 @@
+export const baseUrl = process.env.ALKEMIO_BASE_URL;
+export const urlIdentityRecovery = baseUrl + '/identity/recovery';
+export const urlIdentityLogin = baseUrl + '/identity/login';
+export const urlIdentityRegistration = baseUrl + '/identity/registration'
+export const urlIdentityVerify = baseUrl + '/identity/verify';
+export const urlAdmin = baseUrl + '/admin';
