+++ test/functional-e2e/identity-flows/registration-page-object.ts
