--- test/functional-e2e/identity-flows/verify.e2e-spec.ts
@@ -1,165 +1,0 @@
-import puppeteer from 'puppeteer';
-import VerifyPage from './verify-page-object';
-import CommonActions from '../common/actions';
-import LoginPage from '../authentication/login-page-object';
-import { clearBrowserCookies, getEmails } from '@test/utils/ui.test.helper';
-import RegistrationPage from '../registration/registration-page-object';
-import { uniqueId } from '@test/utils/mutations/create-mutation';
-import {
-  getUser,
-  removeUserMutation,
-} from '@test/functional-api/user-management/user.request.params';
-
-let emailsNumberBefore: number;
-let emailsNumberAfter: number;
-const successAlert =
-  'An email containing a verification link has been sent to the email address you provided.';
-const successMessage = 'Thank you for verifying your email address!';
-
-let userId;
-const emailRegistered = `mail-${uniqueId}@alkem.io`;
-const firstName = 'testFN';
-const lastName = 'testLN';
-const password = process.env.AUTH_TEST_HARNESS_PASSWORD || '';
-const notRegisteredEmail = 'alkemio@test.com';
-const urlIdentityVerify = '/identity/verify';
-const urlIdentityLogin = '/identity/login';
-const commonActions = new CommonActions();
-
-describe('Verification smoke tests', () => {
-  let browser: puppeteer.Browser;
-  let page: puppeteer.Page;
-  beforeAll(async () => {
-    browser = await puppeteer.launch({
-      slowMo: 10,
-      defaultViewport: null,
-      args: ['--window-size=1920,1080'],
-    });
-    page = await browser.newPage();
-    await page.goto(process.env.ALKEMIO_BASE_URL + '/identity/registration', {
-      waitUntil: ['networkidle2', 'domcontentloaded'],
-    });
-    await RegistrationPage.register(
-      page,
-      emailRegistered,
-      password,
-      firstName,
-      lastName
-    );
-  });
-  beforeEach(async () => {
-    let getEmailsData = await getEmails();
-    emailsNumberBefore = getEmailsData[2];
-  });
-
-  afterEach(async () => {
-    await clearBrowserCookies(page);
-  });
-
-  afterAll(async () => {
-    await browser.close();
-    const requestUserData = await getUser(emailRegistered);
-    userId = requestUserData.body.data.user.id;
-    await removeUserMutation(userId);
-  });
-
-  describe('Verify identity flows', () => {
-    test('Verification request from UNauthenticated to registered user', async () => {
-      // Alkemio "verify" page
-      page = await browser.newPage();
-      await page.goto(process.env.ALKEMIO_BASE_URL + `${urlIdentityVerify}`, {
-        waitUntil: ['networkidle2', 'domcontentloaded'],
-      });
-      expect(await VerifyPage.getVerifyPageTitle(page)).toEqual('Verify');
-
-      await VerifyPage.submitVerifyPageForm(page, emailRegistered);
-      expect(await commonActions.getAlertMessage(page)).toEqual(successAlert);
-
-      // Get Url from Email
-      let getEmailsData = await getEmails();
-      let urlFromEmail = getEmailsData[0];
-      if (urlFromEmail === undefined) {
-        throw new Error(`Url from email is missing!`);
-      }
-      emailsNumberAfter = getEmailsData[2];
-
-      // Navigate to the Url
-      await page.goto(urlFromEmail, {
-        waitUntil: ['networkidle2', 'domcontentloaded'],
-      });
-
-      expect(await VerifyPage.getVerifyPageSuccessTitle(page)).toEqual(
-        successMessage
-      );
-      expect(emailsNumberBefore).toEqual(emailsNumberAfter - 1);
-
-      // ToDo: Shouldn't the user be authenticated??
-      // await verifyPage.navigateToUserProfile(pageTwo);
-      // await userProfilePage.verifyUserProfileTitle(page, userFullName);
-    });
-
-    test('Verification request from Authenticated to registered user', async () => {
-      page = await browser.newPage();
-      await page.goto(process.env.ALKEMIO_BASE_URL + `${urlIdentityLogin}`, {
-        waitUntil: ['networkidle2', 'domcontentloaded'],
-      });
-      await LoginPage.login(page, emailRegistered, password);
-
-      // Alkemio "verify" page
-      await page.goto(process.env.ALKEMIO_BASE_URL + `${urlIdentityVerify}`, {
-        waitUntil: ['networkidle2', 'domcontentloaded'],
-      });
-      expect(await VerifyPage.getVerifyPageTitle(page)).toEqual('Verify');
-
-      await VerifyPage.submitVerifyPageForm(page, emailRegistered);
-      expect(await commonActions.getAlertMessage(page)).toEqual(successAlert);
-
-      // Get Url from Email
-      let getEmailsData = await getEmails();
-      let urlFromEmail = getEmailsData[0];
-      if (urlFromEmail === undefined) {
-        throw new Error(`Url from email is missing!`);
-      }
-      emailsNumberAfter = getEmailsData[2];
-
-      // Navigate to the Url
-      await page.goto(urlFromEmail, {
-        waitUntil: ['networkidle2', 'domcontentloaded'],
-      });
-
-      expect(await VerifyPage.getVerifyPageSuccessTitle(page)).toEqual(
-        successMessage
-      );
-      expect(emailsNumberBefore).toEqual(emailsNumberAfter - 1);
-      //ToDo - enable after issue with new link to profile is fixed
-
-      // await VerifyPage.navigateToUserProfile(page);
-      // await UserProfilePage.verifyUserProfileTitle(page, userFullName);
-    });
-
-    test('Verification request from unauthenticated to not registered user', async () => {
-      // Arrange
-      page = await browser.newPage();
-      await page.goto(process.env.ALKEMIO_BASE_URL + `${urlIdentityVerify}`, {
-        waitUntil: ['networkidle2', 'domcontentloaded'],
-      });
-      expect(await VerifyPage.getVerifyPageTitle(page)).toEqual('Verify');
-
-      // Act
-      await VerifyPage.submitVerifyPageForm(page, notRegisteredEmail);
-      expect(await commonActions.getAlertMessage(page)).toEqual(
-        'An email containing a verification link has been sent to the email address you provided.'
-      );
-
-      // Get Url from Email
-      let getEmailsData = await getEmails();
-      emailsNumberAfter = getEmailsData[2];
-
-      // Assert
-      expect(getEmailsData[1]).toContain(
-        'someone asked to verify this email address, but we were unable to find an account for this address'
-      );
-      expect(emailsNumberBefore).toEqual(emailsNumberAfter - 1);
-    });
-  });
-});
